docs/header: Remove register and dashboard from navbar
I think the bar is a bit cluttered, we can put the register link on
the login page, and the dashboard button on the course homepage,
since it is only relevant for current students.

The login button should also be last imo, a lot of websites put it
in the top right.